Fix up windows project files for x64 platform so that builds on 64-bit windows work correctly.
And fix line endings in scan_cta.cu after Andrew's last checkin.  Grrr.

+++ apps/cudpp_testrig/test_radixsort.cu
@@ -119,7 +119,7 @@ int SortSupport<float>::verifySort(float *keysSorted, unsigned int *valuesSorted
 
 template <typename T>
 int radixSortTest(CUDPPHandle plan, CUDPPConfiguration config, size_t *tests, 
-                  unsigned int numTests, unsigned int numElements, unsigned int keybits,
+                  unsigned int numTests, size_t numElements, unsigned int keybits,
                   testrigOptions testOptions, bool quiet)
 {
     int retval = 0;
